requirements.txt
# ============================================================
# Agentic RAG with Docling - Requirements
# ============================================================

# --- Core Document Processing ---
docling>=2.0.0                    # معالجة المستندات الرئيسية
docling-core>=2.0.0               # النماذج الأساسية لـ Docling

# --- LLM & Agents Orchestration ---
langchain>=0.3.0                  # إطار بناء السلاسل
langchain-google-genai>=2.0.0     # تكامل Google Gemini ✅ مجاني
langchain-community>=0.3.0        # أدوات المجتمع
langgraph>=0.2.0                  # بناء State Graph للوكلاء

# --- Embeddings (Google Gemini) ---
google-generativeai>=0.8.0        # مكتبة Gemini الرسمية ✅ مجانية
google-genai>=1.0.0               # Google GenAI SDK (يستخدمه الكود) ✅
tiktoken>=0.8.0                   # عداد الرموز (Tokens)

# --- Database ---
psycopg2-binary>=2.9.9            # اتصال PostgreSQL
pgvector>=0.3.0                   # ملحق المتجهات
sqlalchemy>=2.0.36                # ORM لقواعد البيانات

# --- Metadata & NLP ---
keybert>=0.8.0                    # استخراج الكلمات المفتاحية
sumy>=0.11.0                      # توليد الملخصات
nltk>=3.9.0                       # معالجة اللغة الطبيعية

# --- Evaluation ---
ragas>=0.2.0                      # تقييم أنظمة RAG
langfuse>=2.0.0                   # تتبع الأداء والتكلفة

# --- Utilities ---
python-dotenv>=1.0.0              # إدارة متغيرات البيئة
pydantic>=2.9.0                   # التحقق من البيانات
pydantic-settings>=2.5.0          # إعدادات المشروع
tenacity>=9.0.0                   # إعادة المحاولة عند الفشل
loguru>=0.7.2                     # تسجيل الأحداث (Logging)
rich>=13.9.0                      # طباعة جميلة في الطرفية
httpx>=0.27.0                     # طلبات HTTP

# --- Testing ---
pytest>=8.3.0                     # إطار الاختبارات
pytest-asyncio>=0.24.0            # اختبارات غير متزامنة

# --- UI ---
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9